# internal/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.UserRole:
    enum:
    - PHARMACY
    - WHOLESALER
    - MANUFACTURER
    - ADMIN
    type: string
    x-enum-varnames:
    - UserRolePharmacy
    - UserRoleWholesaler
    - UserRoleManufacturer
    - UserRoleAdmin
...
  response.ErrorEnvelope:
    properties:
      error:
        properties:
          code:
            type: string
          details: {}
          message:
            type: string
        type: object
    type: object
  users.LoginRequest:
    properties:
      identifier:
        type: string
      password:
        type: string
    type: object
  users.RegisterProfile:
    properties:
      address:
        type: string
      city:
        type: string
      country:
        type: string
      license_no:
        type: string
      name:
        type: string
      registration_no:
        type: string
    type: object
  users.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      profile:
        $ref: '#/definitions/users.RegisterProfile'
      role:
        $ref: '#/definitions/model.UserRole'
    type: object
info:
  contact: {}
  description: Backend API for PharmaLink B2B marketplace
  title: PharmaLink API
  version: "1.0"
paths:
...
  /admin/users:
    get:
      parameters:
      - description: role
        in: query
        name: role
        type: string
      - description: status
        in: query
        name: status
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List users (Admin)
      tags:
      - admin
  /admin/users/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      - description: status payload
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Update user status (Admin)
      tags:
      - admin
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with email/phone and password
      parameters:
      - description: login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      summary: Login
      tags:
      - auth
...
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Current user profile
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: refresh payload
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register PHARMACY/WHOLESALER/MANUFACTURER user with profile
      parameters:
      - description: register payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      summary: Register new user
      tags:
      - auth
...
  /rare-requests:
    get:
      parameters:
      - description: status
        in: query
        name: status
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List rare requests
      tags:
      - rare
    post:
      consumes:
      - application/json
      parameters:
      - description: rare request payload
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create rare request (Pharmacy)
      tags:
      - rare
  /rare-requests/{id}/bids:
    post:
      consumes:
      - application/json
      parameters:
      - description: rare request id
        in: path
        name: id
        required: true
        type: string
      - description: rare bid payload
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create rare bid (Wholesaler)
      tags:
      - rare
  /rare-requests/bids/{id}/select:
    post:
      parameters:
      - description: rare bid id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Select rare bid (Pharmacy)
      tags:
      - rare
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"